replace config_version.py by config_version.yaml; move config_loader.py to /pylovo
