+++ classification/config_classification.yaml
@@ -1,5 +1,5 @@
 CLASSIFICATION_VERSION: 1
-VERSION_COMMENT: "set sample with inspected building quality"
+CLASSIFICATION_VERSION_COMMENT: "set sample with inspected building quality"
 CLASSIFICATION_REGION: Bayern
 NO_OF_CLUSTERS_ALLOWED: [3, 4, 5, 6, 7]
 N_SAMPLES: 7  # default value
