--- classification/config_loader.py
@@ -1,40 +1,0 @@
-import yaml
-import os
-
-def load_yaml_config(filepath: str):
-    """Loads a YAML configuration file."""
-    abs_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), filepath)
-    
-    if not os.path.exists(abs_path):
-        raise FileNotFoundError(f"Config file not found: {abs_path}")
-    
-    with open(abs_path, "r", encoding="utf-8") as file:
-        return yaml.safe_load(file)
-
-# Load both configurations with correct paths
-CONFIG_CLASSIFICATION = load_yaml_config("config_classification.yaml")
-CONFIG_CLUSTERING = load_yaml_config("clustering/config_clustering.yaml")
-
-# Assign all variables from CONFIG_CLASSIFICATION
-CLASSIFICATION_VERSION = CONFIG_CLASSIFICATION["CLASSIFICATION_VERSION"]
-VERSION_COMMENT = CONFIG_CLASSIFICATION["VERSION_COMMENT"]
-CLASSIFICATION_REGION = CONFIG_CLASSIFICATION["CLASSIFICATION_REGION"]
-NO_OF_CLUSTERS_ALLOWED = CONFIG_CLASSIFICATION["NO_OF_CLUSTERS_ALLOWED"]
-N_SAMPLES = CONFIG_CLASSIFICATION["N_SAMPLES"]
-REGION_DICT = CONFIG_CLASSIFICATION["REGION_DICT"]
-REGIOSTAR7_DICT = CONFIG_CLASSIFICATION["REGIOSTAR7_DICT"]
-REGIO7_REGIO5_GEM_DICT = CONFIG_CLASSIFICATION["REGIO7_REGIO5_GEM_DICT"]
-
-# Assign all variables from CONFIG_CLUSTERING
-LIST_OF_CLUSTERING_PARAMETERS = CONFIG_CLUSTERING["LIST_OF_CLUSTERING_PARAMETERS"]
-N_CLUSTERS_KMEDOID = CONFIG_CLUSTERING["N_CLUSTERS_KMEDOID"]
-N_CLUSTERS_KMEANS = CONFIG_CLUSTERING["N_CLUSTERS_KMEANS"]
-N_CLUSTERS_GMM = CONFIG_CLUSTERING["N_CLUSTERS_GMM"]
-THRESHOLD_MAX_TRAFO_DIS = CONFIG_CLUSTERING["THRESHOLD_MAX_TRAFO_DIS"]
-THRESHOLD_HOUSEHOLDS_PER_BUILDING = CONFIG_CLUSTERING["THRESHOLD_HOUSEHOLDS_PER_BUILDING"]
-
-# Thresholds for clustering parameters
-THRESHOLD_AVG_TRAFO_DIS = CONFIG_CLUSTERING["THRESHOLD_AVG_TRAFO_DIS"]
-THRESHOLD_NO_HOUSE_CONNECTIONS = CONFIG_CLUSTERING["THRESHOLD_NO_HOUSE_CONNECTIONS"]
-THRESHOLD_VSW_PER_BRANCH = CONFIG_CLUSTERING["THRESHOLD_VSW_PER_BRANCH"]
-THRESHOLD_NO_HOUSEHOLDS = CONFIG_CLUSTERING["THRESHOLD_NO_HOUSEHOLDS"]
