+++ plotting/plotting_classification.py
@@ -13,7 +13,7 @@ from sklearn.mixture import GaussianMixture
 from sklearn.preprocessing import StandardScaler
 from sklearn_extra.cluster import KMedoids
 
-from classification.config_loader import *
+from pylovo.config_loader import *
 from plotting.config_plots import *
