+++ pylovo/GridGenerator.py
@@ -4,8 +4,7 @@ from pathlib import Path
 import pandapower as pp
 from pylovo import pgReaderWriter as pg, utils
 from pylovo.config_data import *
-from pylovo.config_version import *
-from classification.config_loader import CLASSIFICATION_VERSION
+from pylovo.config_loader import *
 
 
 class ResultExistsError(Exception):
