+++ pylovo/config_loader.py
@@ -1,0 +1,56 @@
+import yaml
+import os
+import pandas as pd
+
+def load_yaml_config(filepath: str):
+    """Loads a YAML configuration file."""
+    abs_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), filepath)
+    
+    if not os.path.exists(abs_path):
+        raise FileNotFoundError(f"Config file not found: {abs_path}")
+    
+    with open(abs_path, "r", encoding="utf-8") as file:
+        return yaml.safe_load(file)
+
+# Load all configurations with correct paths
+CONFIG_VERSION = load_yaml_config("config_version.yaml")
+CONFIG_CLASSIFICATION = load_yaml_config("../classification/config_classification.yaml")
+CONFIG_CLUSTERING = load_yaml_config("../classification/clustering/config_clustering.yaml")
+
+# Assign all variables from CONFIG_VERSION
+VERSION_ID = CONFIG_VERSION["VERSION_ID"]
+VERSION_COMMENT = CONFIG_VERSION["VERSION_COMMENT"]
+PLOT_COLOR_DICT = CONFIG_VERSION["PLOT_COLOR_DICT"]
+CONNECTION_AVAILABLE_CABLES = CONFIG_VERSION["CONNECTION_AVAILABLE_CABLES"]
+CABLE_COST_DICT = CONFIG_VERSION["CABLE_COST_DICT"]
+SIM_FACTOR = CONFIG_VERSION["SIM_FACTOR"]
+CONSUMER_CATEGORIES = pd.DataFrame(CONFIG_VERSION["CONSUMER_CATEGORIES"])
+LARGE_COMPONENT_LOWER_BOUND = CONFIG_VERSION["LARGE_COMPONENT_LOWER_BOUND"]
+LARGE_COMPONENT_DIVIDER = CONFIG_VERSION["LARGE_COMPONENT_DIVIDER"]
+VN = CONFIG_VERSION["VN"]
+V_BAND_LOW = CONFIG_VERSION["V_BAND_LOW"]
+V_BAND_HIGH = CONFIG_VERSION["V_BAND_HIGH"]
+
+# Assign all variables from CONFIG_CLASSIFICATION
+CLASSIFICATION_VERSION = CONFIG_CLASSIFICATION["CLASSIFICATION_VERSION"]
+CLASSIFICATION_VERSION_COMMENT = CONFIG_CLASSIFICATION["CLASSIFICATION_VERSION_COMMENT"]
+CLASSIFICATION_REGION = CONFIG_CLASSIFICATION["CLASSIFICATION_REGION"]
+NO_OF_CLUSTERS_ALLOWED = CONFIG_CLASSIFICATION["NO_OF_CLUSTERS_ALLOWED"]
+N_SAMPLES = CONFIG_CLASSIFICATION["N_SAMPLES"]
+REGION_DICT = CONFIG_CLASSIFICATION["REGION_DICT"]
+REGIOSTAR7_DICT = CONFIG_CLASSIFICATION["REGIOSTAR7_DICT"]
+REGIO7_REGIO5_GEM_DICT = CONFIG_CLASSIFICATION["REGIO7_REGIO5_GEM_DICT"]
+
+# Assign all variables from CONFIG_CLUSTERING
+LIST_OF_CLUSTERING_PARAMETERS = CONFIG_CLUSTERING["LIST_OF_CLUSTERING_PARAMETERS"]
+N_CLUSTERS_KMEDOID = CONFIG_CLUSTERING["N_CLUSTERS_KMEDOID"]
+N_CLUSTERS_KMEANS = CONFIG_CLUSTERING["N_CLUSTERS_KMEANS"]
+N_CLUSTERS_GMM = CONFIG_CLUSTERING["N_CLUSTERS_GMM"]
+THRESHOLD_MAX_TRAFO_DIS = CONFIG_CLUSTERING["THRESHOLD_MAX_TRAFO_DIS"]
+THRESHOLD_HOUSEHOLDS_PER_BUILDING = CONFIG_CLUSTERING["THRESHOLD_HOUSEHOLDS_PER_BUILDING"]
+
+# Thresholds for clustering parameters
+THRESHOLD_AVG_TRAFO_DIS = CONFIG_CLUSTERING["THRESHOLD_AVG_TRAFO_DIS"]
+THRESHOLD_NO_HOUSE_CONNECTIONS = CONFIG_CLUSTERING["THRESHOLD_NO_HOUSE_CONNECTIONS"]
+THRESHOLD_VSW_PER_BRANCH = CONFIG_CLUSTERING["THRESHOLD_VSW_PER_BRANCH"]
+THRESHOLD_NO_HOUSEHOLDS = CONFIG_CLUSTERING["THRESHOLD_NO_HOUSEHOLDS"]
