+++ pylovo/config_version.yaml
@@ -1,0 +1,60 @@
+# Change version if you changed any grid parameters or queries (e.g. transformer query)
+VERSION_ID: 1.0
+
+# state here which parameters, queries are used to compute the grids with current VERSION_ID
+VERSION_COMMENT: base version
+
+PLOT_COLOR_DICT:
+  100: gold
+  160: orange
+  200: darkorange
+  250: peru
+  315: orangered
+  400: red
+  515: crimson
+  500: crimson
+  630: purple
+  715: blue
+  800: darkblue
+  1260: black
+
+CONNECTION_AVAILABLE_CABLES:
+  - NAYY 4x120 SE
+  - NAYY 4x95 SE
+  - NAYY 4x50 SE
+  - NYY 4x16 SE
+  - NYY 4x35 SE
+  - NYY 4x70 SE
+
+# PARAMETERS
+CABLE_COST_DICT:
+  NAYY 4x185 SE: 33
+  NAYY 4x150 SE: 24
+  NAYY 4x120 SE: 19
+  NAYY 4x95 SE: 16
+  NAYY 4x50 SE: 11
+  NYY 4x16 SE: 7
+  NYY 4x35 SE: 16
+  NYY 4x95 SE: 37
+  NYY 4x70 SE: 28
+
+# installed_power * (sim_factor + (1 - sim_factor) * (load_count ** (-3 / 4)))
+SIM_FACTOR: {Residential: 0.07, Public: 0.6, Commercial: 0.5}
+
+# format: consumer_category_id: [definition, peak_load, yearly_consumption, peak_load_per_m2, yearly_consumption_per_m2, sim_factor]
+CONSUMER_CATEGORIES:
+  - {consumer_category_id: 1, definition: Commercial, peak_load: null, yearly_consumption: null, peak_load_per_m2: 79.00, yearly_consumption_per_m2: 245.70, sim_factor: 0.50}
+  - {consumer_category_id: 2, definition: Public, peak_load: null, yearly_consumption: null, peak_load_per_m2: 29.00, yearly_consumption_per_m2: 155.70, sim_factor: 0.60}
+  - {consumer_category_id: 9, definition: SFH, peak_load: &PEAK_LOAD_HOUSEHOLD 30.00, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+  - {consumer_category_id: 10, definition: MFH, peak_load: *PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+  - {consumer_category_id: 11, definition: TH, peak_load: *PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+  - {consumer_category_id: 12, definition: AB, peak_load: *PEAK_LOAD_HOUSEHOLD, yearly_consumption: 0.00, peak_load_per_m2: null, yearly_consumption_per_m2: null, sim_factor: 0.07}
+
+# Larger components are divided by k-means clustering in roughly LARGE_COMPONENT_DIVIDER sizes
+LARGE_COMPONENT_LOWER_BOUND: 2000
+LARGE_COMPONENT_DIVIDER: 1000
+
+# Voltage properties
+VN: 400  # V
+V_BAND_LOW: 0.95  # +-5%
+V_BAND_HIGH: 1.05
